+++ web/apps/photos/src/services/logout.ts
@@ -3,8 +3,8 @@ import {
     logoutClearStateAgain,
 } from "@/accounts/services/logout";
 import log from "@/base/log";
+import { downloadManager } from "@/gallery/services/download";
 import { resetUploadState } from "@/gallery/upload";
-import DownloadManager from "@/new/photos/services/download";
 import { logoutML, terminateMLWorker } from "@/new/photos/services/ml";
 import { logoutSearch } from "@/new/photos/services/search";
 import { logoutSettings } from "@/new/photos/services/settings";
@@ -30,7 +30,7 @@ export const photosLogout = async () => {
     try {
         await terminateMLWorker();
     } catch (e) {
-        ignoreError("ml/worker", e);
+        ignoreError("ML/worker", e);
     }
 
     // - Remote logout and clear state
@@ -44,31 +44,31 @@ export const photosLogout = async () => {
     try {
         logoutSettings();
     } catch (e) {
-        ignoreError("settings", e);
+        ignoreError("Settings", e);
     }
 
     try {
         logoutUserDetails();
     } catch (e) {
-        ignoreError("userDetails", e);
+        ignoreError("User details", e);
     }
 
     try {
         resetUploadState();
     } catch (e) {
-        ignoreError("upload", e);
+        ignoreError("Upload", e);
     }
 
     try {
-        DownloadManager.logout();
+        downloadManager.logout();
     } catch (e) {
-        ignoreError("download", e);
+        ignoreError("Download", e);
     }
 
     try {
         logoutSearch();
     } catch (e) {
-        ignoreError("search", e);
+        ignoreError("Search", e);
     }
 
     // - Desktop
@@ -84,13 +84,13 @@ export const photosLogout = async () => {
         try {
             exportService.disableContinuousExport();
         } catch (e) {
-            ignoreError("export", e);
+            ignoreError("Export", e);
         }
 
         try {
             await electron.logout();
         } catch (e) {
-            ignoreError("electron", e);
+            ignoreError("Electron", e);
         }
     }
